backend/requirements.txt
asgiref==3.3.4
certifi==2024.2.2
charset-normalizer==3.3.2
Deprecated==1.2.14
distro==1.9.0
Django==3.2.3
idna==3.7
netaddr==1.2.1
Pillow==8.3.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2021.1
requests==2.31.0
sqlparse==0.4.1
urllib3==2.2.1
wrapt==1.16.0
yookassa==3.1.0
